fix(passport): change favicon to client id icon when present so wallet and browser shows branded 3rd party logo
